+++ changelog.md
@@ -2,7 +2,14 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9248]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9249]: https://www.github.com/atecoder/dn.vi.ivi
+
+## [8.1.9249] - 2025-04-28
+- Device Tsp Script:
+  - Add node auto run toggle methods.
+  - Add node save method.
+- Firmware Manager:
+  - fix saving script with auto run.
 
 ## [8.1.9248] - 2025-04-27
 - Device Tsp Scrip Namespace